# backend/public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Car Ordering Service Backend",
        "description": "Backend API documentation for the Car Ordering Service project.",
        "contact": {
            "name": "Otavie Okuoyo",
            "email": "otavieokuoyo@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:5468",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/sender/": {
            "post": {
                "tags": [
                    "Sender"
                ],
                "summary": "Create a New Sender",
                "description": "Endpoint to create a new sender. It requires the sender",
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "example": "any"
                                    },
                                    "email": {
                                        "example": "any"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/driver/": {
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "Create a New Driver",
                "description": "Endpoint to create a new driver. Upon creation, the driver is marked as available.",
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/order/": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Create a New Order",
                "description": "Endpoint to create a new order request. Upon successful creation, the order is initially set to pending status. After 1 minute, if not accepted by a driver, it will be automatically rejected.",
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "location": {
                                        "example": "any"
                                    },
                                    "destination": {
                                        "example": "any"
                                    },
                                    "price": {
                                        "example": "any"
                                    },
                                    "senderId": {
                                        "example": "any"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/order/{orderId}/accept": {
            "post": {
                "tags": [
                    "Order/{orderId}/accept"
                ],
                "summary": "Accept Order",
                "description": "Endpoint for a driver to accept an order. Upon successful acceptance, the order status will be updated to accepted.",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "driverId": {
                                        "example": "any"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
            "Order": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "object"
                    },
                    "properties": {
                        "type": "object",
                        "properties": {
                            "location": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "The current location where the order originates from."
                                    }
                                }
                            },
                            "destination": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "The desired destination (address, landmark) of the car ride."
                                    }
                                }
                            },
                            "price": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "number"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "The price of the order, usually in the currency specified by the system."
                                    }
                                }
                            },
                            "senderId": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "The unique identifier of the sender who placed the order."
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "Sender": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "object"
                    },
                    "properties": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Name of the sender"
                                    }
                                }
                            },
                            "email": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Email of the sender."
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "Driver": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "object"
                    },
                    "properties": {
                        "type": "object",
                        "properties": {
                            "isAvailable": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "boolean"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Indicates whether the driver is available."
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}